Update normative.py
Rewritten large chunks thanks to @morganjwilliams feedback:
 - removed many superfluous np.where statements
- added bit more commenting (still working on this)
- called on periodic_table for molecular masses
- spotted a couple of minor bugs

I am finding this improves the output a little bit. However, there are still discrepancies between the output of this code and that of Verma's software - the culprits being nepheline, perovskite and (less so) magnetite. I have a bit more time I want to spend on this based on Morgan's comments, hopefully by the end of the week, otherwise early next week.

Let me know if I've messed up on Git anywhere, still getting used to this!