Create template for QAP ternary diagram
This sets up the classifier for a QAP diagram. Most of the details are copied from the existing USDA soil classifier, but modified.

The polygons are in the right locations, but need to still be tested with some sample data to ensure that the classification is working as expected.

The unusual presentation of the standard QAP ternary compared to a normal one probably needs to be called out in the docs as well, which I have not done yet.